making ci pass

+++ packaging/gen-packages.sh
@@ -6,13 +6,16 @@ if [[ -z "${VERSION}" ]]; then
     exit 2
 fi
 
+if [[ -z "${ARCH}" ]]; then
+    echo "\$ARCH enviroment variable undefined"
+    exit 2
+fi
+
 rm -rf pkg-build
 
-echo "Building k6... for v: $VERSION and ${VESION}"
+echo "Building k6..."
 CGO_ENABLED=0 GOARCH=$ARCH go build -a -ldflags '-s -w' -o /tmp/k6
 echo "Done!"
-VERSION=0.21.0
-ARCH=amd64
 
 mkdir -p dist
 VERBOSE=* go-bin-$1 generate --file packaging/$1.json -a $ARCH --version $VERSION -o dist/k6-v$VERSION-$ARCH.deb
